Fix 3 bugs + security: dealer API 500, OTP emoji crash, open redirect
- Fix ProductImage.path -> file_path in dealer service (500 on /api/dealer/products)
- Fix emoji print() crash on Windows cp1252 in sms.py and settings.py
- Add open redirect protection for next_url in auth routes
- Purity column: Integer -> Numeric(4,1) for 999.9 support

+++ modules/catalog/models.py
@@ -61,7 +61,7 @@ class Product(Base):
     id = Column(Integer, primary_key=True, index=True)
     name = Column(String, nullable=False)
     weight = Column(Numeric(10, 3), nullable=False)           # گرم
-    purity = Column(Integer, default=750, nullable=False)      # عیار (750 = 18K)
+    purity = Column(Numeric(4, 1), default=750, nullable=False)  # عیار (750, 995, 999.9)
     wage = Column(Numeric(5, 2), default=0, nullable=False)     # اجرت ساخت (درصدی)
     is_wage_percent = Column(Boolean, default=True, nullable=False)  # همیشه True برای شمش
     design = Column(String, nullable=True)
